advanced-api-project/pyproject.toml
[tool.poetry]
name = "advanced-api-project"
version = "0.1.0"
description = ""
authors = ["j1ere <jeremywambua106@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
